# docs/swagger.yaml
basePath: /
definitions:
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  model.TreatementRequest:
    description: Treatment request information
    properties:
      issues:
        example: Back pain
        type: string
      next_visit:
        example: "2025-01-22"
        type: string
      patient_code:
        example: J001
        type: string
      remarks:
        example: Patient showed improvement
        type: string
      therapist_id:
        example: 1
        type: integer
      treatment:
        example:
        - Massage therapy
        - Exercise
        items:
          type: string
        type: array
      treatment_date:
        example: "2025-01-15"
        type: string
    type: object
  model.Treatment:
    description: Treatment information
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        example: 1
        type: integer
      issues:
        example: Back pain
        type: string
      next_visit:
        example: "2025-01-22"
        type: string
      patient_code:
        example: J001
        type: string
      remarks:
        example: Patient showed improvement
        type: string
      therapist_id:
        example: 1
        type: integer
      treatment:
        example: Massage therapy,Exercise
        type: string
      treatment_date:
        example: "2025-01-15"
        type: string
      updatedAt:
        type: string
    type: object
  util.APIResponse:
    properties:
      data: {}
      error:
        type: string
      msg:
        type: string
      success:
        type: boolean
    type: object
host: localhost:19091
info:
  contact:
    email: support@ariebrainware.com
    name: Arie Brainware
  description: |-
    REST API for managing patient data, treatments, and therapy sessions
    This API provides endpoints for patient management, disease tracking, treatment records, and therapist management.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: LTT Backend API
  version: "1.0"
paths:
...
  /treatment:
    get:
      consumes:
      - application/json
      description: Get a paginated list of treatments with optional filtering by therapist,
        keyword, and date
      parameters:
      - description: Limit number of results
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Filter by therapist ID
        in: query
        name: therapist_id
        type: integer
      - description: Search keyword for patient name or patient code
        in: query
        name: keyword
        type: string
      - description: Filter by specific date (YYYY-MM-DD format)
        in: query
        name: group_by_date
        type: string
      - description: Filter by logged-in therapist
        in: query
        name: filter_by_therapist
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Treatments fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Invalid request or session error
          schema:
            $ref: '#/definitions/util.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/util.APIResponse'
      security:
      - BearerAuth: []
      - SessionToken: []
      summary: List all treatments
      tags:
      - Treatment
    post:
      consumes:
      - application/json
      description: Add a new treatment record
      parameters:
      - description: Treatment information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.TreatementRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Treatment created successfully
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Invalid request or duplicate treatment
          schema:
            $ref: '#/definitions/util.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/util.APIResponse'
      security:
      - BearerAuth: []
      - SessionToken: []
      summary: Create a new treatment
      tags:
      - Treatment
  /treatment/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a treatment record by ID
      parameters:
      - description: Treatment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Treatment deleted successfully
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Treatment not found
          schema:
            $ref: '#/definitions/util.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/util.APIResponse'
      security:
      - BearerAuth: []
      - SessionToken: []
      summary: Delete a treatment
      tags:
      - Treatment
    patch:
      consumes:
      - application/json
      description: Update an existing treatment record
      parameters:
      - description: Treatment ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated treatment information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.Treatment'
      produces:
      - application/json
      responses:
        "200":
          description: Treatment updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Treatment'
              type: object
        "400":
          description: Invalid request or treatment not found
          schema:
            $ref: '#/definitions/util.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/util.APIResponse'
      security:
      - BearerAuth: []
      - SessionToken: []
      summary: Update treatment information
      tags:
      - Treatment
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  SessionToken:
    description: Session token for authenticated requests
    in: header
    name: session-token
    type: apiKey
swagger: "2.0"